requirements.txt
requests
beautifulsoup4
selenium
webdriver-manager
undetected-chromedriver
pandas
plotly
kaleido
pycountry
bibtexparser
numpy
scikit-learn
scipy
matplotlib
networkx
wordcloud
pillow

# Django y REST API
Django>=5.2
djangorestframework>=3.14
django-cors-headers>=4.3

# Playwright para scraping headless
playwright>=1.40
gunicorn>=21.2
